feat(streaming): migrate TopNExecutor to ExecutorV2
