+++ src/stream/src/executor_v2/error.rs
@@ -36,6 +36,9 @@ pub enum StreamExecutorError {
     #[error("aggregate state error {0}")]
     AggStateError(RwError),
 
+    #[error("top n state error {0}")]
+    TopNStateError(RwError),
+
     #[error("channel `{0}` closed")]
     ChannelClosed(String),
 }
@@ -55,6 +58,10 @@ impl StreamExecutorError {
 
     pub fn agg_state_error(error: impl Into<RwError>) -> TracedStreamExecutorError {
         Self::AggStateError(error.into()).into()
+    }
+
+    pub fn top_n_state_error(error: impl Into<RwError>) -> TracedStreamExecutorError {
+        Self::TopNStateError(error.into()).into()
     }
 
     pub fn channel_closed(name: impl Into<String>) -> TracedStreamExecutorError {
