+++ src/stream/src/executor_v2/mod.rs
@@ -39,6 +39,8 @@ pub mod receiver;
 mod simple;
 #[cfg(test)]
 mod test_utils;
+mod top_n;
+mod top_n_executor;
 mod v1_compat;
 
 pub use batch_query::BatchQueryExecutor;
@@ -50,6 +52,7 @@ pub use local_simple_agg::LocalSimpleAggExecutor;
 pub use merge::MergeExecutor;
 pub use mview::*;
 pub(crate) use simple::{SimpleExecutor, SimpleExecutorWrapper};
+pub use top_n::TopNExecutor;
 pub use v1_compat::StreamExecutorV1;
 
 pub type BoxedExecutor = Box<dyn Executor>;
