+++ src/stream/src/executor_v2/v1_compat.rs
@@ -21,7 +21,7 @@ use risingwave_common::catalog::ColumnId;
 pub use risingwave_common::catalog::Schema;
 use risingwave_common::error::{ErrorCode, Result, RwError};
 use risingwave_common::hash::HashKey;
-use risingwave_common::util::sort_util::OrderPair;
+use risingwave_common::util::sort_util::{OrderPair, OrderType};
 use risingwave_expr::expr::BoxedExpression;
 use risingwave_storage::table::cell_based_table::CellBasedTable;
 use risingwave_storage::{Keyspace, StateStore};
@@ -35,6 +35,7 @@ use super::{
 pub use super::{BoxedMessageStream, ExecutorV1, Message, PkIndices, PkIndicesRef};
 use crate::executor::AggCall;
 use crate::executor_v2::global_simple_agg::SimpleAggExecutor;
+use crate::executor_v2::top_n::TopNExecutor;
 use crate::task::FinishCreateMviewNotifier;
 
 /// The struct wraps a [`BoxedMessageStream`] and implements the interface of [`ExecutorV1`].
@@ -273,5 +274,34 @@ impl<S: StateStore> BatchQueryExecutor<S> {
         };
 
         Self::new(table, Self::DEFAULT_BATCH_SIZE, info, key_indices)
+    }
+}
+
+impl<S: StateStore> TopNExecutor<S> {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new_from_v1(
+        input: Box<dyn ExecutorV1>,
+        pk_order_types: Vec<OrderType>,
+        offset_and_limit: (usize, Option<usize>),
+        pk_indices: PkIndices,
+        keyspace: Keyspace<S>,
+        cache_size: Option<usize>,
+        total_count: (usize, usize, usize),
+        executor_id: u64,
+        _op_info: String,
+        key_indices: Vec<usize>,
+    ) -> Result<Self> {
+        let input = Box::new(ExecutorV1AsV2(input));
+        Self::new(
+            input,
+            pk_order_types,
+            offset_and_limit,
+            pk_indices,
+            keyspace,
+            cache_size,
+            total_count,
+            executor_id,
+            key_indices,
+        )
     }
 }
